image-processing-api/requirements.txt
fastapi
uvicorn
gunicorn
requests
Pillow
python-dotenv